// authority-portal-frontend/src/app/pages/catalog-page/policy-editor/model/policy-expression-mapped.ts
/*
 * Data Space Portal
 * Copyright (C) 2025 sovity GmbH
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU Affero General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU Affero General Public License for more details.
 *
 * You should have received a copy of the GNU Affero General Public License
 * along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
import {UiPolicyLiteral} from '@sovity.de/authority-portal-client';
import {PolicyMultiExpressionConfig} from './policy-multi-expressions';
import {PolicyOperatorConfig} from './policy-operators';
import {PolicyVerbConfig} from './policy-verbs';

export interface PolicyExpressionMapped {
  type: 'CONSTRAINT' | 'MULTI' | 'EMPTY';

  multiExpression?: PolicyMultiExpressionConfig;
  expressions?: PolicyExpressionMapped[];

  verb?: PolicyVerbConfig;
  operator?: PolicyOperatorConfig;
  valueRaw?: UiPolicyLiteral;
  valueJson?: string;
  displayValue?: string;
}